updated swagger spec for API 1.1.0
the previous swagger spec was broken because of a basic code generation. this needs to be maintained more carefully and not reliant upon generated outputs. @akutz fixed this part: You must have a  tag for a parameter of type . It was added wherever it was omitted, setting the schema to a variant type instead. It was omitted in a lot of requests as well as one response. Within the lookups, I (@kacole2) changed the lookup query reuirement from FALSE to TRUE. For some reason the code would fail to execute unless it was set to TRUE. There were some memory pointers happening and it will have to be looked at once again. To get a better idea of how this is all generated using go-swagger, checkout the gorackhd repo and follow the directions listed there. https://github.com/emccode/gorackhd

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>